+++ src/resources/containers/instances/instance.ts
@@ -51,7 +51,7 @@ export type InstanceState =
  */
 export type ReadyState = "active" | "purge" | "hibernate";
 
-export type InstanceEvent = "first_boot" | "started";
+export type InstanceEvent = "first_started";
 export type InstanceQuery = QueryParams<
   keyof InstanceIncludes,
   keyof InstanceMetas
@@ -190,8 +190,10 @@ export async function remove(
 
 export async function removeMultiple(
   params: StandardParams<InstanceQuery> & {
-    instances_ids: ResourceId[];
     containerId: ResourceId;
+    value: {
+      instances_ids: ResourceId[];
+    };
   },
 ) {
   return Request.deleteRequest<CreatedTask<any>>({