+++ src/resources/hubs/activity.ts
@@ -95,6 +95,7 @@ export type EventType =
   | "image.create"
   | "image.delete"
   | "image.task.delete"
+
   // containers
   | "container.create"
   | "container.start"
@@ -112,13 +113,15 @@ export type EventType =
   | "container.task.scale"
   | "container.delete"
   | "container.task.delete"
-  | "container.instances.create"
   // instances
   | "container.instance.error"
   | "container.instance.sftp.login"
   | "container.instance.migration.start"
   | "container.instance.migration.revert"
   | "container.instance.delete"
+  | "container.instances.delete"
+  | "container.instances.create"
+
   // dns zones
   | "dns.zone.task.verify"
   | "dns.zone.task.delete"