+++ src/resources/hubs/providers/equinix-metal.ts
@@ -1,4 +1,4 @@
-export interface PacketProvider {
+export interface EquinixMetalProvider {
   api_key: string;
   project: string;
 }