+++ src/resources/surveys/survey.ts
@@ -7,9 +7,11 @@ export type SurveySingle = SingleDoc<Survey>;
 export type SendResponseParams = Response;
 
 export interface Survey extends Resource {
+  hub_scoped: ResourceId;
   identifier: string;
   active: boolean;
   title: string;
+  reward?: Reward;
   sections: Section[];
 }
 
@@ -27,6 +29,10 @@ export interface Question {
   dropdown?: Record<string, string>;
   required?: boolean;
   conditional_questions?: Record<string, Question[]>;
+}
+
+export interface Reward {
+  credit?: number;
 }
 
 export async function getSingle(