+++ CHANGELOG.md
@@ -9,6 +9,24 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- `container.instances.delete` activity
+
+### Changed
+
+- `containerId` => `container_id` for `removeMultiple` instances
+
+### Breaking Changes
+
+- Packet provider changes to Equinix Metal
+- `PacketProvider` -> `EquinixMetalProvider`
+- `ProviderIdentifier = "packet" | ...` -> `ProviderIdentifier = "equinix_metal" | ...`
+
+## [1.9.2]
+
+### Fixed
+
+### Added
+
 - Survey structures
 - Survey link(s)
 - Initialize task action for environments